+++ src/PlantBiophysics.jl
@@ -7,10 +7,7 @@ import PlantSimEngine: Status
 import PlantMeteo
 import PlantMeteo: Weather, AbstractAtmosphere
 
-# For IO:
-import YAML
 import CSV
-import OrderedCollections: OrderedDict
 import Dates
 
 import DataFrames.DataFrame # For convenience transformations
@@ -40,10 +37,8 @@ include("simulation.jl")
 include("processes/γ_star.jl")
 
 # Light interception
-include("processes/light/Ignore.jl")
 include("processes/light/Beer.jl")
 include("processes/light/Beer_shortwave.jl")
-include("processes/light/Translucent.jl")
 
 # Photosynthesis related files:
 include("processes/photosynthesis/constantA.jl")
@@ -68,7 +63,6 @@ include("processes/energy/Monteith.jl")
 
 # File IO
 include("io/read_file.jl")
-include("io/read_model.jl")
 include("io/read_walz.jl")
 include("io/read_licor6400.jl")
 include("io/read_licor6800.jl")
@@ -80,9 +74,7 @@ include("fitting/fit_FvCB.jl")
 include("fitting/fit_Medlyn.jl")
 include("fitting/fit_Beer.jl")
 
-# File IO:
-export read_model
-export is_model
+# File IO
 export read_walz, read_licor6400, read_licor6800, read_ess_dive, read_ciras4
 
 # Conversions
@@ -105,13 +97,6 @@ export net_longwave_radiation
 export Monteith       # a struct to hold the values for the model of Monteith and Unsworth (2013)
 export latent_heat, sensible_heat
 
-# structure for light interception
-export Translucent
-export LightIgnore
-export OpticalProperties
-export σ
-export AbstractLight_InterceptionModel
-
 # Photosynthesis
 export AbstractPhotosynthesisModel
 export ConstantA, ConstantAGs
@@ -121,16 +106,13 @@ export FvcbRaw # Parameters for the original Farquhar et al. (1980) model
 
 # Conductances
 export AbstractStomatal_ConductanceModel
-export gbh_to_gbw
 export gbₕ_free
 export gbₕ_forced
 export Medlyn, Tuzet, ConstantGs
-export ConstantGs
 
 # Model helpers
 export get_km, Γ_star, arrhenius, get_J, gs_closure, get_Cᵢⱼ, get_Cᵢᵥ, get_Dₕ
-export Fvcb_net_assimiliation
-export get_process, get_model, instantiate
+export Fvcb_net_assimilation, Fvcb_net_assimiliation
 export leaf_scene
 
 end